Elaborate concepts and procedures (#1)
* Concepts: Grammar cleanup

An "installation" obviously has stuff "installed", we should define
things a little more specifically.

* Concepts: Fix typos

* Concept: Improve grammar for the ubuntu-devel branch definition

* Concepts: Copyedit how to access merge proposal status

Put the directions in order of propose merge, then merge status.

This also fixes some clumsy grammar of the Either/Or phrasing.

* README: Mention the doc will explain git-ubuntu

There are a number of other packaging tutorials out there, so to
differentiate ourselves emphasize from the start that we'll tie in with
use of git ubuntu.

* Concepts: Explain deb-src

This follows logically from the earlier explanations of binary
vs. source packages.  The reader of this document is, by nature,
interested in source packages, so it is pertinent to show how to enable
them.

* Concepts: Update git-ubuntu installation directions

The Setup document directs the user to install git-ubuntu from the snap
package's edge channel.  For consistency, provide the same directions
here.  Currently, only the edge channel gets regular testing, so is
still the best approach for getting git-ubuntu.

* README: Note that this is not a policy document

* Concepts:  Explain how Debian, Ubuntu, Canonical, and Launchpad relate

Expand the introduction a bit to outline the principal groups involved
in package management, and how they're all related.  No need for deep
detail, just define the terms.

* Concepts: Explain briefly about orig tarballs, .dsc files, and changelogs.

* Add mention of launchpad-gm-scripts for bug triage

* README: Note this is not intended to be comprehensive documentation

* Improve explanation of changelog entry wording

* MergeProposal: Add recommendations for preparing for sponsorship requests

* PackageBuilding: build-source is replaced by build --source

* Setup: lxc aliases

* Setup: Add directions for adding user to groups

* Setup: Note gpg key strength consideration esp. for old keys

* Setup: Additional ~/.gitconfig suggestions

* Setup: Elaborate a bit on quilt

* Setup: Define what dput is briefly, and what the suggested settings do

* Setup: Define what lxd is, briefly

* Concepts: Define release cadences and SRUs

* Explain sponsorship briefly

* PackageFixing: Add section for fixed-upstream bugs

The guide already covered issues fixed in Ubuntu and fixed in Debian, so
add the case of issues fixed upstream for completeness.

I expect this can be expanded a fair bit, but for now point out uscan
and general tips for finding stuff upstream.

* CommittingChanges: Add guidelines on writing commit messages

N.b. These tips are taken largely verbatim from a Xenial-era server team
onboarding guide.

* PackageFixing: Detail the SRU process

N.b. These tips are taken largely verbatim from a Xenial-era server team
onboarding guide.