PackageFixing:  Illustrate finding bugs using Debian's git repo
The current example for finding fixes in Debian is a limited in that it
won't find fixes not already included in Ubuntu somewhere.  Instead suggest
a more general approach of using Debian's git repositories and/or salsa.

(git-ubuntu itself is changing how it shows the changelog info, and may
also be gaining tracking of Debian branches in the future, so this
section is worth reworking again in the future.)