PackageFixing: Select branch, lxc, and ppa names consistently in the example
The existing examples are derived from an assortment of different
packaging tasks, resulting in a diverse variety of lxc, branch, and ppa
names.  This is a bit hard to follow for students (and makes the docs a
bit tricky to maintain).

Instead, try to maintain focus on the simple 'at' example, and select
names using a consistent pattern.  Some of the latter examples are still
needed since they different cover aspects of packaging.