Put some words into the 'Raise issues' section
We have a pretty well established process for our triage.
Just the tail end with the pings/mails is done, but not "defined"
within these documents.

This adds a description of our mailing list summaries and the pings
that we sometimes do so that we can discuss and formally agree
on the behavior.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>