BugTriage.md: use multi line bullet lists
Currently our layout is like:

* Support Requests. Sometimes users report problems that are really
  just a misunderstanding of how to use their system.
  - action

That is not very readable when first scrolling over it to find the
category one is looking for. OTOH making the details part of the bullet
list seems like it would be an action. And adding another indent level
wastes space and makes the list mroe noisy.

Instead use multi line bullet lists starting with the category and
continuing on the next line with the details about that category.
The above example would then be:

* Support Requests.
  Sometimes users report problems that are really just a misunderstanding
  of how to use their system.
  - action

To do that in markdown one needs to end a bullet point with two white spaces
before the line break.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>