BugTriage.md: update tagging and housekeeping
We have said that we want to try the weekly housekeeping session
as well as dropping server-next tag for a while and make it part of the
documentation once we are happy with it. It seems to work fine so let us
put it into the handbook.

In addition that will help others to run the meeting as this will list
the commands to generate bug lists as well as where to find old bug
tracking lists.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>